Use the behavior mock in other tests
