+++ tests/conftest.py
@@ -22,6 +22,20 @@ else:
 
 @pytest.fixture
 def ak_behavior(mocker):
+    "Patch the awkward behavior dictionary in `util_types`"
     the_dict = {}
     mocker.patch("layered_edm.util_types.ak.behavior", the_dict)
     return the_dict
+
+
+@pytest.fixture
+def simple_ds():
+    import awkward as ak
+
+    return ak.Array(
+        [
+            [{"x": 1}, {"x": 2}, {"x": 3}],
+            [],
+            [{"x": 4}, {"x": 5}],
+        ]
+    )
